% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_f}
\alias{fmt_f}
\title{Format an F-test result}
\usage{
fmt_f(
  f_value,
  df1,
  df2,
  p,
  eta_sq = NULL,
  partial = TRUE,
  digits = 2,
  markup = TRUE
)
}
\arguments{
\item{f_value}{F statistic}

\item{df1}{Numerator degrees of freedom}

\item{df2}{Denominator degrees of freedom}

\item{p}{P-value}

\item{eta_sq}{Eta-squared or partial eta-squared}

\item{partial}{Whether eta_sq is partial}

\item{digits}{Decimal places}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format an F-test result
}
