% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_mean_sd}
\alias{fmt_mean_sd}
\title{Format mean and standard deviation}
\usage{
fmt_mean_sd(mean, sd, digits = 2, markup = TRUE)
}
\arguments{
\item{mean}{Mean value}

\item{sd}{Standard deviation}

\item{digits}{Decimal places}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format mean and standard deviation
}
