% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_mean_se}
\alias{fmt_mean_se}
\title{Format mean and standard error}
\usage{
fmt_mean_se(mean, se, digits = 2, markup = TRUE)
}
\arguments{
\item{mean}{Mean value}

\item{se}{Standard error}

\item{digits}{Decimal places}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format mean and standard error
}
