% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_p}
\alias{fmt_p}
\title{Format a p-value in APA style}
\usage{
fmt_p(p, digits = 3, include_p = TRUE, markup = TRUE)
}
\arguments{
\item{p}{Numeric p-value}

\item{digits}{Decimal places (default 3)}

\item{include_p}{Include "p = " prefix}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format a p-value in APA style
}
\examples{
fmt_p(0.043)
fmt_p(0.0003)
}
