% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{format_decimal}
\alias{format_decimal}
\title{Format a decimal number}
\usage{
format_decimal(x, digits = 2, leading_zero = TRUE)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}

\item{leading_zero}{Include leading zero for values between -1 and 1}
}
\value{
Character string
}
\description{
Format a decimal number
}
