% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_funcs.R
\name{sync_clear}
\alias{sync_clear}
\title{Clear all data from the live server}
\usage{
sync_clear()
}
\value{
Invisible NULL.
}
\description{
Deletes all tracked statistics and tables from the global StatSync state,
ensuring that future exports or syncs start with a blank slate.
}
