% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sync_correlation_table}
\alias{sync_correlation_table}
\title{Create a correlation matrix table}
\usage{
sync_correlation_table(
  data,
  vars = NULL,
  method = "pearson",
  digits = 2,
  caption = NULL
)
}
\description{
Create a correlation matrix table
}
