% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{sync_export_all}
\alias{sync_export_all}
\title{Quick export from RMarkdown chunk}
\usage{
sync_export_all(envir = parent.frame(), file = NULL)
}
\arguments{
\item{envir}{Environment to search for statsync objects}

\item{file}{Output file}
}
\description{
Designed to be called within an RMarkdown document. Automatically
collects all statsync objects in the environment and exports.
}
