% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_funcs.R
\name{sync_stop}
\alias{sync_stop}
\title{Stop the active StatSync server}
\usage{
sync_stop()
}
\value{
Invisible NULL.
}
\description{
Shuts down the local HTTP server if one is currently running via
\code{sync_serve()}.
}
