% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{sync_switch}
\alias{sync_switch}
\title{Switch active StatSync project}
\usage{
sync_switch(project_name)
}
\arguments{
\item{project_name}{The name of the project to switch to}
}
\value{
Invisible project name
}
\description{
Switches the current active project, automatically loading any previously
saved data for the new project from disk.
}
