% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_chi}
\alias{fmt_chi}
\title{Format a chi-square test result}
\usage{
fmt_chi(chi_value, df, p, v = NULL, digits = 2, markup = TRUE)
}
\arguments{
\item{chi_value}{Chi-square statistic}

\item{df}{Degrees of freedom}

\item{p}{P-value}

\item{v}{Cramér's V (optional)}

\item{digits}{Decimal places}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format a chi-square test result
}
