% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format a confidence interval}
\usage{
fmt_ci(
  lower,
  upper,
  digits = 2,
  level = 95,
  bracket_type = "square",
  markup = TRUE
)
}
\arguments{
\item{lower}{Lower bound}

\item{upper}{Upper bound}

\item{digits}{Decimal places}

\item{level}{Confidence level (for labeling)}

\item{bracket_type}{"square" or "round"}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format a confidence interval
}
