% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_apa.R
\name{fmt_t}
\alias{fmt_t}
\title{Format a t-test result (standalone t-test)}
\usage{
fmt_t(t_value, df, p, d = NULL, digits = 2, markup = TRUE)
}
\arguments{
\item{t_value}{t statistic}

\item{df}{Degrees of freedom}

\item{p}{P-value}

\item{d}{Cohen's d (optional)}

\item{digits}{Decimal places for statistics}

\item{markup}{Include italic markup tags}
}
\value{
Character string
}
\description{
Format a t-test result (standalone t-test)
}
